requirements.txt
aiogram==3.17.0
python-dotenv==1.0.0
aiofiles>=23.2.1
Pillow==10.1.0
psutil==5.9.5
aiohttp>=3.9.0,<3.14
pytz==2024.1
uvicorn==0.25.0
fastapi==0.108.0
redis==5.0.1
prometheus-client==0.19.0
requests==2.31.0
python-json-logger==2.0.7

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
